+++ src/client/components/evaluate/detail/Competitiveness.tsx
@@ -7,6 +7,7 @@ import PVIDisplay from "../../PVIDisplay";
 import { formatPvi, computeRowFill, calculatePVI } from "../../../functions";
 import { checkPlanScoreAPI } from "../../../api";
 import { IProject, PlanScoreAPIResponse } from "../../../../shared/entities";
+import CompetitivenessChart from "./CompetitivenessChart";
 
 const style: ThemeUIStyleObject = {
   table: {
@@ -86,10 +87,11 @@ const CompetitivenessMetricDetail = ({
     <Box>
       <Heading as="h2" sx={{ variant: "text.h5", mt: 4 }}>
         Partisan Voting Index (PVI):
-        <span sx={{ color: metric.party?.color || "#000" }}>
+        <span sx={{ color: metric.party?.color || "#000", ml: "10px", mb: "10px" }}>
           {formatPvi(metric.party, metric.value)}
         </span>
       </Heading>
+      <CompetitivenessChart geojson={geojson} metric={metric} />
       <Styled.table sx={style.table}>
         <thead>
           <Styled.tr>